s3blaster: miscellaneous improvements
1. Filling objects:
Objects are filled with a range and a number of parallel requests.

2. Control list of uploaded objects
An uploaded object' keys make a change in the `storedKeys` array:
    storedKeys[bucket][object size][object key] := true
If `workOnCurrObjs` flag is true, get and delete operations will work on
only objects whose key is stored in `storedKeys`.

3. Better handle for get/delete operations
In cases of get/delete operations, if there is no key, i.e. no object,
- if sequential simulation: set count = threshold in order to go to next simul
- otherwise, go next (mixed) operation

4. Plotter
Added header in graph files containing stats results